refactor: move and rename service provider class
