+++ give.php
@@ -54,7 +54,6 @@ use Give\DonorDashboards\ServiceProvider as DonorDashboardsServiceProvider;
 use Give\DonorDashboards\Tabs\TabsRegister;
 use Give\Donors\Repositories\DonorRepositoryProxy;
 use Give\Donors\ServiceProvider as DonorsServiceProvider;
-use Give\FeatureFlags\FeatureFlagsServiceProvider;
 use Give\Form\LegacyConsumer\ServiceProvider as FormLegacyConsumerServiceProvider;
 use Give\Form\Templates;
 use Give\FormBuilder\ServiceProvider as FormBuilderServiceProvider;
@@ -243,7 +242,7 @@ final class Give
         Give\FormTaxonomies\ServiceProvider::class,
         Give\DonationSpam\ServiceProvider::class,
         Give\Campaigns\ServiceProvider::class,
-        FeatureFlagsServiceProvider::class,
+        Give\FeatureFlags\OptionBasedFormEditor\ServiceProvider::class,
     ];
 
     /**
